moving resolve_globbing function in separate file
